# frontend/openapi.snapshot.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "AI-Powered Multi-Agent Career Opportunity Platform",
    "version": "0.1.0"
  },
  "paths": {
...
    "/jobs/ingest-url": {
      "post": {
        "tags": [
          "job-ingestion"
        ],
        "summary": "Ingest Job Url Route",
        "description": "Submit a job posting URL for extraction.\n\nNote (flagged contract gap \u2014 see this agent's final report):\napi-contracts.md lists `user_id exists` as a validation rule producing\n404 `NOT_FOUND`, but dependency-graph.md documents zero runtime API\ncalls from Job Ingestion Service to User Service, so this handler has\nno documented way to verify `user_id` exists and does not attempt to.\n`user_id` is now token-derived rather than client-supplied (see\n`infrastructure.auth`), which closes the \"client can claim any\nuser_id\" gap but not this one \u2014 a valid token's user_id still isn't\nchecked for existence in the `users` table.",
        "operationId": "ingest_job_url_route_jobs_ingest_url_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/IngestJobUrlRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/jobs/{job_id}": {
      "get": {
        "tags": [
          "job-ingestion"
        ],
        "summary": "Get Job Route",
        "description": "Read one job's canonical detail. See api-contracts.md#job-ingestion-service\nfor why this lives here rather than on Job Matching Service (the `jobs`\ntable's actual owner/writer, shared between this service and Job\nDiscovery Service \u2014 see database-ownership.md#jobs).",
        "operationId": "get_job_route_jobs__job_id__get",
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Job Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/job-sources": {
      "get": {
        "tags": [
          "job-discovery"
        ],
        "summary": "List Job Sources",
        "operationId": "list_job_sources_job_sources_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/JobSourceResponse"
                  },
                  "type": "array",
                  "title": "Response List Job Sources Job Sources Get"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      },
      "post": {
        "tags": [
          "job-discovery"
        ],
        "summary": "Create Job Source",
        "operationId": "create_job_source_job_sources_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateJobSourceRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobSourceResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
...
    "/jobs/{job_id}/contacts": {
      "get": {
        "tags": [
          "contacts"
        ],
        "summary": "Get Job Contacts",
        "description": "See gap 1 in this module's docstring \u2014 no job-existence check is\npossible from this component.",
        "operationId": "get_job_contacts_jobs__job_id__contacts_get",
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Job Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ContactResponse"
                  },
                  "title": "Response Get Job Contacts Jobs  Job Id  Contacts Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/jobs/{job_id}/contacts/search": {
      "post": {
        "tags": [
          "contacts"
        ],
        "summary": "Trigger Contact Search",
        "description": "Thin producer only \u2014 publishes `ContactsRequestedEvent` onto the\nsame `contacts.requested` command topic Job Matching Service uses\nautomatically (kafka-topics.md's design-decision note grants this\ncomponent producer rights here for exactly this manual re-trigger,\n\"not a self-call\"). The result arrives later via `contacts.found`,\nhandled by this component's own `contacts.requested` consumer just\nlike the automatic path \u2014 see gap 2 in this module's docstring for why\nthe request body carries `company`/`title` (`user_id` is now\ntoken-derived rather than a third additive body field).",
        "operationId": "trigger_contact_search_jobs__job_id__contacts_search_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Job Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TriggerContactSearchRequest"
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TriggerContactSearchResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ContactResponse": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "full_name": {
            "type": "string",
            "title": "Full Name"
          },
          "headline": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Headline"
          },
          "company": {
            "type": "string",
            "title": "Company"
          },
          "contact_type": {
            "$ref": "#/components/schemas/ContactType"
          },
          "profile_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Profile Url"
          },
          "relevance_score": {
            "type": "number",
            "title": "Relevance Score"
          },
          "status": {
            "$ref": "#/components/schemas/ContactStatus"
          }
        },
        "type": "object",
        "required": [
          "id",
          "full_name",
          "company",
          "contact_type",
          "relevance_score",
          "status"
        ],
        "title": "ContactResponse"
      },
      "ContactStatus": {
        "type": "string",
        "enum": [
          "DISCOVERED",
          "RANKED",
          "ARCHIVED"
        ],
        "title": "ContactStatus",
        "description": "Owner: Contact Discovery Service. See\nstate-machines.md#contact-lifecycle-contactstatus--contactstatus."
      },
      "ContactType": {
        "type": "string",
        "enum": [
          "PRACTITIONER",
          "TEAM_LEAD",
          "HIRING_MANAGER",
          "RECRUITER",
          "EXECUTIVE",
          "DEPARTMENT_LEADER",
          "OTHER"
        ],
        "title": "ContactType",
        "description": "Owner: Contact Discovery Service.\n\nDeliberately profession-generic \u2014 captures functional tier only\n(practitioner vs. lead vs. hiring manager vs. recruiter vs. executive).\nA profession-specific title (e.g. \"Senior Mechanical Engineer\") belongs\nin `Contact.headline` as free text, never as a new enum member here. See\nshared-types.md#shared-enums."
      },
      "CreateJobSourceRequest": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          },
          "type": {
            "$ref": "#/components/schemas/JobSourceType"
          },
          "query_config": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Query Config"
          },
          "enabled": {
            "type": "boolean",
            "title": "Enabled"
          }
        },
        "type": "object",
        "required": [
          "name",
          "type",
          "enabled"
        ],
        "title": "CreateJobSourceRequest"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "IngestJobUrlRequest": {
        "properties": {
          "url": {
            "type": "string",
            "title": "Url"
          }
        },
        "type": "object",
        "required": [
          "url"
        ],
        "title": "IngestJobUrlRequest"
      },
...
      "JobProcessingStatus": {
        "type": "string",
        "enum": [
          "PENDING",
          "NORMALIZED",
          "MATCHED",
          "FAILED"
        ],
        "title": "JobProcessingStatus",
        "description": "Owner: Job Ingestion / Job Discovery Service."
      },
      "JobResponse": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "user_id": {
            "type": "string",
            "format": "uuid",
            "title": "User Id"
          },
          "company": {
            "type": "string",
            "title": "Company"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "location": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Location"
          },
          "description": {
            "type": "string",
            "title": "Description"
          },
          "extracted_skills": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Extracted Skills",
            "default": []
          },
          "experience_required": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Experience Required"
          },
          "source_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Url"
          },
          "processing_status": {
            "$ref": "#/components/schemas/JobProcessingStatus"
          },
          "discovered_at": {
            "type": "string",
            "format": "date-time",
            "title": "Discovered At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "user_id",
          "company",
          "title",
          "description",
          "processing_status",
          "discovered_at"
        ],
        "title": "JobResponse"
      },
      "JobSourceResponse": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "user_id": {
            "type": "string",
            "format": "uuid",
            "title": "User Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "type": {
            "$ref": "#/components/schemas/JobSourceType"
          },
          "query_config": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Query Config"
          },
          "enabled": {
            "type": "boolean",
            "title": "Enabled"
          },
          "last_run_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Run At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "user_id",
          "name",
          "type",
          "enabled"
        ],
        "title": "JobSourceResponse"
      },
      "JobSourceType": {
        "type": "string",
        "enum": [
          "MANUAL_URL",
          "LINKEDIN",
          "INDEED",
          "COMPANY_CAREERS_PAGE",
          "OTHER"
        ],
        "title": "JobSourceType",
        "description": "Owner: Job Discovery Service. Extensible \u2014 adding a source is additive."
      },
...
      "TriggerContactSearchRequest": {
        "properties": {
          "company": {
            "type": "string",
            "title": "Company"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "location": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Location"
          }
        },
        "type": "object",
        "required": [
          "company",
          "title"
        ],
        "title": "TriggerContactSearchRequest",
        "description": "See this module's docstring \u2014 an additive gap-resolution type, not\npart of api-contracts.md's documented (path-param-only) input."
      },
      "TriggerContactSearchResponse": {
        "properties": {
          "job_id": {
            "type": "string",
            "format": "uuid",
            "title": "Job Id"
          },
          "requested_at": {
            "type": "string",
            "format": "date-time",
            "title": "Requested At"
          }
        },
        "type": "object",
        "required": [
          "job_id",
          "requested_at"
        ],
        "title": "TriggerContactSearchResponse"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}